fixed warning

+++ src/ConvexPolygonMesh.cpp
@@ -36,7 +36,7 @@ Triangle ConvexPolygonMesh::Triangles::generate() const {
 
 	Triangle triangle{};
 
-	const int verticesPerRing = mMesh->mSegments * mMesh->mVertices.size();
+	const int verticesPerRing = int(mMesh->mSegments * mMesh->mVertices.size());
 	const int delta = mRingIndex * verticesPerRing + 1;
 
 	const int n1 = mSideIndex * mMesh->mSegments + mSegmentIndex;
